refactor: change user handler package and imports in server.go
