+++ articlehub-api/internal/handler/user_handler/user.go
@@ -1,4 +1,4 @@
-package handler
+package user_handler
 
 import (
 	"bytes"
